azure-ml-deployment/Python/requirements.txt
azure-storage-blob==0.37.1
azure
azureml-sdk